requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0

# Video Processing
moviepy==1.0.3
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80

# Audio Analysis
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# AI/ML
google-generativeai==0.3.2
mediapipe==0.10.9

# Utilities
Pillow==10.2.0
requests==2.31.0
tqdm==4.66.1
pydantic==2.5.3
pydantic-settings==2.1.0
